pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-api-hub"
version = "0.1.0"
description = "Unified Python interface for multiple AI model APIs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "api", "llm", "vlm", "image-generation", "audio", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/model-api-hub"
Repository = "https://github.com/yourusername/model-api-hub"
Issues = "https://github.com/yourusername/model-api-hub/issues"

[project.scripts]
model-api-hub = "model_api_hub.cli:main"
modelhub = "model_api_hub.cli:main"

[tool.setuptools]
packages = ["model_api_hub"]

[tool.setuptools.package-data]
model_api_hub = ["*.yaml", "*.example"]
